requirements.txt
plotly-express 
pandas
seaborn
imbalanced-learn
joblib